change routes to router and remove keyword.rs
