+++ server/src/router/mod.rs
@@ -1,6 +1,14 @@
+use super::Config;
+use axum::{
+    http::StatusCode,
+    routing::{get, get_service},
+    Router,
+};
+use axum_extra::routing::SpaRouter;
+use tower_http::services::ServeFile;
+
 use axum::{
     extract::{Path, Query as AxumQuery},
-    http::StatusCode,
     response::IntoResponse,
     Extension,
 };
@@ -9,7 +17,34 @@ use std::collections::HashMap;
 
 use crate::queries;
 
-use super::processor::SignalProcessor;
+use processor::SignalProcessor;
+pub mod processor;
+
+pub trait Build {
+    fn build_v1(cfg: &Config) -> Router;
+}
+
+impl Build for Router {
+    fn build_v1(cfg: &Config) -> Router {
+        let spa = SpaRouter::new("/assets", format!("{}/assets", &cfg.client_path))
+            .index_file("index.html");
+
+        let react_router_fallback =
+            get_service(ServeFile::new(format!("{}/index.html", &cfg.client_path))).handle_error(
+                |error: std::io::Error| async move {
+                    (
+                        StatusCode::INTERNAL_SERVER_ERROR,
+                        format!("Internal Server Error: {}", error),
+                    )
+                },
+            );
+
+        Router::new()
+            .merge(spa)
+            .route("/api/v1/:keyword", get(handler))
+            .fallback(react_router_fallback)
+    }
+}
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub enum QueryRequestFilter {
